add laser test for entity

+++ aparriott.h
@@ -27,8 +27,8 @@ class Entity {
 	float pos[2];
     float curve[2];
 	unsigned char color[3];
-	int hp;	// how much hp he starts out with
-	int damage;	// how much damage he does if collides with toaster
+	float hp;	// how much hp he starts out with
+	float damage;	// how much damage he does if collides with toaster
 	int point;
 	/*void set_color(unsigned char col[3]) {
 		memcpy(color,  col, sizeof(unsigned char) * 3);
